// types/supabase.ts
export interface definitions {
// ...
  brands: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    uuid: string;
    website?: string;
    industry?: string;
    location?: string;
    founded?: string;
    size?: string;
    description?: string;
    name?: string;
    thumbnail?: string;
  };
// ...
  posts: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    title?: string;
    description?: string;
    type?: string;
    views?: number;
    options?: string;
    uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `users.uuid`.<fk table='users' column='uuid'/>
     */
    user: string;
    image?: string;
    url?: string;
    /**
     * Note:
     * This is a Foreign Key to `topics.uuid`.<fk table='topics' column='uuid'/>
     */
    topic?: string;
  };
  posts_related_brands: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.uuid`.<fk table='posts' column='uuid'/>
     */
    post?: string;
    /**
     * Note:
     * This is a Foreign Key to `brands.uuid`.<fk table='brands' column='uuid'/>
     */
    brand?: string;
  };
// ...
  topics: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    name: string;
    uuid: string;
  };
// ...
}